refactor: move updateQuarterText para common de hypertension backend
